refactor: utils, constants, values regrouped
