+++ libs/constants.ts
@@ -1,2 +1,141 @@
+export const imgSrcReplaceReg = new RegExp(
+  /src=[\\"\']?([^>\\"\']+)[\\"\']?[^>]*>/g,
+);
+
 export const THEME_META_IMAGE =
   'https://res.cloudinary.com/dolziw8fv/image/upload/v1674413434/DALL_E_2023-01-23_03.46.28_-_a_sceintific_photo_of_the_most_nostalgic_sundog_during_sunset_on_ocean_provia_400_rillrw.png"';
+export const THEME_LOGO_IMAGE =
+  "https://res.cloudinary.com/dolziw8fv/image/upload/v1663058798/logo_ktizgo.jpg";
+export const DEFAULT_REACT_IMAGE =
+  "https://res.cloudinary.com/dolziw8fv/image/upload/v1661838248/handinthesky_by7xjb.jpg";
+export const DEFAULT_NEXT_IMAGE =
+  "https://res.cloudinary.com/dolziw8fv/image/upload/v1661669755/zlc6jlp1p6ilr8rhzvi5.png";
+export const DEFAULT_TS_IMAGE =
+  "https://res.cloudinary.com/dolziw8fv/image/upload/v1661840879/typescript_preset_ioudd0.png";
+export const DEFAULT_JS_IMAGE =
+  "https://res.cloudinary.com/dolziw8fv/image/upload/v1662480649/javascript_preset_f07j3f.png";
+export const DEFAULT_ETC_IMAGE =
+  "https://res.cloudinary.com/dolziw8fv/image/upload/v1661838134/component_pji40f.png";
+export const DEFAULT_RECAP_IMAGE =
+  "https://res.cloudinary.com/dolziw8fv/image/upload/v1661975868/recap-preset_ssksj1.png";
+export const DEFAULT_ERROR_IMAGE =
+  "https://res.cloudinary.com/dolziw8fv/image/upload/v1661975868/error-preset_s0chxc.png";
+export const DEFAULT_IMAGE =
+  "https://res.cloudinary.com/dolziw8fv/image/upload/v1674554187/rkvskmgscq88ku3kzfks.png";
+export const DEFAULT_SKY =
+  "https://res.cloudinary.com/dolziw8fv/image/upload/v1674554187/rkvskmgscq88ku3kzfks.png";
+const GALLERY_2022_IMAGES = [
+  {
+    url: "https://res.cloudinary.com/dolziw8fv/image/upload/v1661975838/gallery/v1/gallery-v1--56_b1qhqt.jpg",
+    title: "",
+    createdAt: "2014",
+    width: "680px",
+    height: "1024px",
+  },
+  {
+    url: "https://res.cloudinary.com/dolziw8fv/image/upload/v1661975837/gallery/v1/gallery-v1--48_ylniyt.jpg",
+    title: "",
+    createdAt: "2018",
+    width: "680pxpx",
+    height: "1024px",
+  },
+  {
+    url: "https://res.cloudinary.com/dolziw8fv/image/upload/v1661975836/gallery/v1/gallery-v1--42_alerpg.jpg",
+    title: "",
+    createdAt: "2018",
+    width: "680px",
+    height: "1024px",
+  },
+  {
+    url: "https://res.cloudinary.com/dolziw8fv/image/upload/v1662479712/gallery/v1/gallery-v1--58_pafn9u.jpg",
+    title: "",
+    createdAt: "2022",
+    width: "680px",
+    height: "1024px",
+  },
+  {
+    url: "https://res.cloudinary.com/dolziw8fv/image/upload/v1661975838/gallery/v1/gallery-v1--53_q4b9hr.jpg",
+    title: "unfold",
+    createdAt: "2021",
+    width: "1024px",
+    height: "680px",
+  },
+  {
+    url: "https://res.cloudinary.com/dolziw8fv/image/upload/v1661975837/gallery/v1/gallery-v1--44_vd2be7.jpg",
+    title: "",
+    createdAt: "2018",
+    width: "680px",
+    height: "1024px",
+  },
+  {
+    url: "https://res.cloudinary.com/dolziw8fv/image/upload/v1661975836/gallery/v1/gallery-v1--45_zihs6p.jpg",
+    title: "Iguazu",
+    createdAt: "2018",
+    width: "680px",
+    height: "1024px",
+  },
+  {
+    url: "https://res.cloudinary.com/dolziw8fv/image/upload/v1662050042/gallery-v1--31_ud9i4l.jpg",
+    title: "",
+    createdAt: "2018",
+    width: "800px",
+    height: "800px",
+  },
+  {
+    url: "https://res.cloudinary.com/dolziw8fv/image/upload/v1661975836/gallery/v1/gallery-v1--41_efzhkx.jpg",
+    title: "",
+    createdAt: "2018",
+    width: "800px",
+    height: "800px",
+  },
+  {
+    url: "https://res.cloudinary.com/dolziw8fv/image/upload/v1661975835/gallery/v1/gallery-v1--34_nlztmy.jpg",
+    title: "curves",
+    createdAt: "2016",
+    width: "730px",
+    height: "1024px",
+  },
+  {
+    url: "https://res.cloudinary.com/dolziw8fv/image/upload/v1661975835/gallery/v1/gallery-v1--37_mahgnu.jpg",
+    title: "",
+    createdAt: "2018",
+    width: "800px",
+    height: "800px",
+  },
+  {
+    url: "https://res.cloudinary.com/dolziw8fv/image/upload/v1661975835/gallery/v1/gallery-v1--29_gmgn79.jpg",
+    title: "",
+    createdAt: "2016",
+    width: "730px",
+    height: "1024px",
+  },
+  {
+    url: "https://res.cloudinary.com/dolziw8fv/image/upload/v1661975836/gallery/v1/gallery-v1--46_opobxq.jpg",
+    title: "",
+    createdAt: "2019",
+    width: "680px",
+    height: "1024px",
+  },
+  {
+    url: "https://res.cloudinary.com/dolziw8fv/image/upload/v1661975838/gallery/v1/gallery-v1--52_qrvivh.jpg",
+    title: "",
+    createdAt: "2021",
+    width: "680px",
+    height: "1024px",
+  },
+  {
+    url: "https://res.cloudinary.com/dolziw8fv/image/upload/v1662050056/preset-pic-8_lwrnoh.jpg",
+    title: "",
+    createdAt: "2018",
+    width: "680px",
+    height: "1024px",
+  },
+  {
+    url: "https://res.cloudinary.com/dolziw8fv/image/upload/v1661975837/gallery/v1/gallery-v1--51_zm5vox.jpg",
+    title: "",
+    createdAt: "2021",
+    width: "2024px",
+    height: "1400px",
+  },
+];
+export { GALLERY_2022_IMAGES };
