--- libs/galleryImages.ts
@@ -1,116 +1,0 @@
-const galleryImagesV1 = [
-  {
-    url: "https://res.cloudinary.com/dolziw8fv/image/upload/v1661975838/gallery/v1/gallery-v1--56_b1qhqt.jpg",
-    title: "",
-    createdAt: "2014",
-    width: "680px",
-    height: "1024px",
-  },
-  {
-    url: "https://res.cloudinary.com/dolziw8fv/image/upload/v1661975837/gallery/v1/gallery-v1--48_ylniyt.jpg",
-    title: "",
-    createdAt: "2018",
-    width: "680pxpx",
-    height: "1024px",
-  },
-  {
-    url: "https://res.cloudinary.com/dolziw8fv/image/upload/v1661975836/gallery/v1/gallery-v1--42_alerpg.jpg",
-    title: "",
-    createdAt: "2018",
-    width: "680px",
-    height: "1024px",
-  },
-  {
-    url: "https://res.cloudinary.com/dolziw8fv/image/upload/v1662479712/gallery/v1/gallery-v1--58_pafn9u.jpg",
-    title: "",
-    createdAt: "2022",
-    width: "680px",
-    height: "1024px",
-  },
-  {
-    url: "https://res.cloudinary.com/dolziw8fv/image/upload/v1661975838/gallery/v1/gallery-v1--53_q4b9hr.jpg",
-    title: "unfold",
-    createdAt: "2021",
-    width: "1024px",
-    height: "680px",
-  },
-  {
-    url: "https://res.cloudinary.com/dolziw8fv/image/upload/v1661975837/gallery/v1/gallery-v1--44_vd2be7.jpg",
-    title: "",
-    createdAt: "2018",
-    width: "680px",
-    height: "1024px",
-  },
-  {
-    url: "https://res.cloudinary.com/dolziw8fv/image/upload/v1661975836/gallery/v1/gallery-v1--45_zihs6p.jpg",
-    title: "Iguazu",
-    createdAt: "2018",
-    width: "680px",
-    height: "1024px",
-  },
-  {
-    url: "https://res.cloudinary.com/dolziw8fv/image/upload/v1662050042/gallery-v1--31_ud9i4l.jpg",
-    title: "",
-    createdAt: "2018",
-    width: "800px",
-    height: "800px",
-  },
-  {
-    url: "https://res.cloudinary.com/dolziw8fv/image/upload/v1661975836/gallery/v1/gallery-v1--41_efzhkx.jpg",
-    title: "",
-    createdAt: "2018",
-    width: "800px",
-    height: "800px",
-  },
-  {
-    url: "https://res.cloudinary.com/dolziw8fv/image/upload/v1661975835/gallery/v1/gallery-v1--34_nlztmy.jpg",
-    title: "curves",
-    createdAt: "2016",
-    width: "730px",
-    height: "1024px",
-  },
-  {
-    url: "https://res.cloudinary.com/dolziw8fv/image/upload/v1661975835/gallery/v1/gallery-v1--37_mahgnu.jpg",
-    title: "",
-    createdAt: "2018",
-    width: "800px",
-    height: "800px",
-  },
-  {
-    url: "https://res.cloudinary.com/dolziw8fv/image/upload/v1661975835/gallery/v1/gallery-v1--29_gmgn79.jpg",
-    title: "",
-    createdAt: "2016",
-    width: "730px",
-    height: "1024px",
-  },
-  {
-    url: "https://res.cloudinary.com/dolziw8fv/image/upload/v1661975836/gallery/v1/gallery-v1--46_opobxq.jpg",
-    title: "",
-    createdAt: "2019",
-    width: "680px",
-    height: "1024px",
-  },
-  {
-    url: "https://res.cloudinary.com/dolziw8fv/image/upload/v1661975838/gallery/v1/gallery-v1--52_qrvivh.jpg",
-    title: "",
-    createdAt: "2021",
-    width: "680px",
-    height: "1024px",
-  },
-  {
-    url: "https://res.cloudinary.com/dolziw8fv/image/upload/v1662050056/preset-pic-8_lwrnoh.jpg",
-    title: "",
-    createdAt: "2018",
-    width: "680px",
-    height: "1024px",
-  },
-  {
-    url: "https://res.cloudinary.com/dolziw8fv/image/upload/v1661975837/gallery/v1/gallery-v1--51_zm5vox.jpg",
-    title: "",
-    createdAt: "2021",
-    width: "2024px",
-    height: "1400px",
-  },
-];
-
-export { galleryImagesV1 };
