--- libs/post/handleDeleteBlogPost.ts
@@ -1,7 +1,0 @@
-const handleDeleteBlogPost = (url: string) => {
-  fetch(url, {
-    method: "DELETE",
-  });
-};
-
-export default handleDeleteBlogPost;
