--- libs/utils/dayJs.ts
@@ -1,5 +1,0 @@
-const dayJs = (date: string): string => {
-  return date.slice(0, 10).replaceAll("-", "/");
-};
-
-export default dayJs;
