--- libs/utils/quillFormat.tsx
@@ -1,35 +1,0 @@
-const formats = [
-  "header",
-  "font",
-  "size",
-  "bold",
-  "italic",
-  "underline",
-  "strike",
-  "align",
-  "blockquote",
-  "code-block",
-  "list",
-  "bullet",
-  "indent",
-  "background",
-  "color",
-  "link",
-  "image",
-  "video",
-  "width",
-];
-
-const toolbarOptions = [
-  ["link", "image", "video"],
-  [{ header: [1, 2, 3, false] }],
-  ["bold", "italic", "underline", "strike"],
-  ["blockquote", "code-block"],
-  [{ list: "ordered" }, { list: "bullet" }],
-  [{ color: [] }, { background: [] }],
-  [{ size: ["small", false, "large", "huge"] }],
-  [{ align: [] }],
-  [{ font: [] }],
-];
-
-export { formats, toolbarOptions };
