--- libs/utils/regExp.ts
@@ -1,3 +1,0 @@
-const imgSrcReplaceReg = new RegExp(/src=[\\"\']?([^>\\"\']+)[\\"\']?[^>]*>/g);
-
-export { imgSrcReplaceReg };
